refactor particles classes
